requirements.txt
llama-index==0.14.2
llama-index-llms-ollama==0.7.3
ollama==0.5.4
python-dotenv==1.0.0
requests==2.32.3
